requirements.txt
web3>=6.0.0
py-solc-x
flask
python-dotenv
reportlab